+++ rules/DowngradePhp70/Rector/Declare_/DowngradeStrictTypeDeclarationRector.php
@@ -2,58 +2,38 @@
 
 declare(strict_types=1);
 
-namespace Rector\DowngradePhp70\Rector\Class_;
+namespace Rector\DowngradePhp70\Rector\Declare_;
 
 use PhpParser\Node;
-use PhpParser\Node\Stmt\Class_;
+use PhpParser\Node\Stmt\Declare_;
 use Rector\Core\Rector\AbstractRector;
 use Symplify\RuleDocGenerator\ValueObject\CodeSample\CodeSample;
 use Symplify\RuleDocGenerator\ValueObject\RuleDefinition;
 
 /**
- * @see \Rector\Tests\DowngradePhp70\Rector\Class_\DowngradeAnonymousClassRector\DowngradeAnonymousClassRectorTest
+ * @see \Rector\Tests\DowngradePhp70\Rector\Declare_\DowngradeStrictTypeDeclarationRector\DowngradeStrictTypeDeclarationRectorTest
  */
-final class DowngradeAnonymousClassRector extends AbstractRector
+final class DowngradeStrictTypeDeclarationRector extends AbstractRector
 {
     /**
      * @return array<class-string<Node>>
      */
     public function getNodeTypes(): array
     {
-        return [Class_::class];
+        return [Declare_::class];
     }
 
     public function getRuleDefinition(): RuleDefinition
     {
         return new RuleDefinition(
-            'Remove anonymous class',
+            'Remove the declare(strict_types=1)',
             [
                 new CodeSample(
                     <<<'CODE_SAMPLE'
-class SomeClass
-{
-    public function run()
-    {
-        return new class {
-            public function execute()
-            {
-            }
-        };
-    }
-}
+declare(strict_types=1);
 CODE_SAMPLE
                     ,
                     <<<'CODE_SAMPLE'
-class Anonymous
-{
-}
-class SomeClass
-{
-    public function run()
-    {
-        return new Anonymous();
-    }
-}
 CODE_SAMPLE
                 ),
             ]
@@ -61,10 +41,28 @@ CODE_SAMPLE
     }
 
     /**
-     * @param Class_ $node
+     * @param Declare_ $node
      */
     public function refactor(Node $node): ?Node
     {
+        if ($this->shouldSkip($node)) {
+            return null;
+        }
+
+        $this->removeNode($node);
         return $node;
+    }
+
+    private function shouldSkip(Declare_ $declare): bool
+    {
+        $declares = $declare->declares;
+
+        foreach ($declares as $declare) {
+            if ($this->isName($declare->key, 'strict_types')) {
+                return false;
+            }
+        }
+
+        return true;
     }
 }
